Fix flag bug

+++ CryptoPkg/Driver/Crypto.c
@@ -4930,7 +4930,7 @@ CryptoServiceEcDhComputeKey (
   IN OUT  VOID         *EcContext,
   IN      CONST UINT8  *PeerPublic,
   IN      UINTN        PeerPublicSize,
-  IN      CONST UINT8  *CompressFlag,
+  IN      CONST INT32  *CompressFlag,
   OUT     UINT8        *Key,
   IN OUT  UINTN        *KeySize
   )
