+++ CryptoPkg/Library/BaseCryptLibNull/Ec/CryptEcNull.c
@@ -485,7 +485,7 @@ EcDhComputeKey (
   IN OUT  VOID         *EcContext,
   IN      CONST UINT8  *PeerPublic,
   IN      UINTN        PeerPublicSize,
-  IN      CONST UINT8  *CompressFlag,
+  IN      CONST INT32  *CompressFlag,
   OUT     UINT8        *Key,
   IN OUT  UINTN        *KeySize
   )
